fix replacement of config files

+++ MateCat-Trusty/MateCatApache/run.sh
@@ -29,11 +29,11 @@ fi
 
 # MateCat
 MATECAT_VERSION=$(fgrep '=' ./inc/version.ini | awk '{print $3}')
-cp /tmp/config.ini ./inc/
-cp /tmp/node_config.ini ./nodejs/config.ini
-cp /tmp/oauth_config.ini ./inc/
-cp /tmp/Error_Mail_List.ini ./inc/
-cp /tmp/task_manager_config.ini ./daemons/
+[[ ! -f './inc/config.ini' ]] && cp /tmp/config.ini ./inc/
+[[ ! -f './nodejs/config.ini' ]] && cp /tmp/node_config.ini ./nodejs/config.ini
+[[ ! -f './inc/oauth_config.ini' ]] && cp /tmp/oauth_config.ini ./inc/
+[[ ! -f './inc/Error_Mail_List.ini' ]] &&  cp /tmp/Error_Mail_List.ini ./inc/
+[[ ! -f './daemons/task_manager_config.ini' ]]  && cp /tmp/task_manager_config.ini ./daemons/
 
 sed -ri -e "s/X.X.X/${MATECAT_VERSION}/g" ./inc/config.ini
 sed -ri -e "s/_SMTP_HOST_/${SMTP_HOST}/g" ./inc/config.ini
